removed recvq since it will not be used modified pcm buffer to be an 8bit pointer instead of 16bit and cast as 16bit when needed for the sake of readability
git-svn-id: http://svn.mangler.org/mangler/trunk@59 f654edcb-a5fb-4f28-a0e7-68c327795dae

+++ libventrilo3/ventrilo3.h
@@ -203,7 +203,7 @@ struct _v3_event {
     struct {
         uint32_t length;
         uint32_t rate;
-        uint16_t sample[32768];
+        uint8_t  sample[32768];
     } pcm;
 
     v3_event *next;
